Cleanup files and add them to package.xml
