+++ src/ext/backtrace.h
@@ -1,7 +1,7 @@
-#ifndef DD_DEBUG_BACKTRACE_H
-#define DD_DEBUG_BACKTRACE_H
+#ifndef DD_BACKTRACE_H
+#define DD_BACKTRACE_H
 #define MAX_STACK_SIZE 1024
 void ddtrace_backtrace_handler(int sig);
 void ddtrace_install_backtrace_handler();
 
-#endif  // DD_DEBUG_H
+#endif  // DD_BACKTRACE_H
