+++ src/ext/ddtrace.c
@@ -13,7 +13,7 @@
 #include "compat_zend_string.h"
 #include "ddtrace.h"
 #include "debug.h"
-#include "debug_backtrace.h"
+#include "backtrace.h"
 #include "dispatch.h"
 #include "dispatch_compat.h"
 #include "request_hooks.h"
@@ -297,17 +297,3 @@ ZEND_GET_MODULE(ddtrace)
 ZEND_TSRMLS_CACHE_DEFINE();
 #endif
 #endif
-
-void ddtrace_log_errf(const char *format, ...)
-{
-	va_list args;
-	char *buffer;
-	size_t size;
-
-	va_start(args, format);
-	size = vspprintf(&buffer, 0, format, args);
-	ddtrace_log_err(buffer);
-
-    efree(buffer);
-	va_end(args);
-}
